+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - Add `has_withdrawal_period` to `Product` model
 - Prevent to a user to create an order for a product without withdrawal period
   if the user has not waived his withdrawal right
+- Display `has_waived_withdrawal_right` in back office Order view
 
 ## [2.9.2] - 2024-10-24